requirements.txt
streamlit
pandas
numpy==1.23.5
matplotlib
seaborn
scikit-learn
nltk
spacy==3.4.4
textblob
python-docx
google-generativeai
openai
anthropic
Pillow
thinc==8.1.10
torch==2.3.1+cpu
openpyxl